requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aioschedule==0.5.2
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
APScheduler==3.10.4
apscheduler-di==0.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
black==24.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
emoji==2.10.1
environs==10.3.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
gspread==6.0.2
gspread-formatting==1.1.2
gspread_asyncio==2.0.0
httplib2==0.22.0
idna==3.6
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.0
multidict==6.0.5
mypy-extensions==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
protobuf==4.25.3
psycopg==3.1.18
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rodi==2.0.6
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.27
StrEnum==0.4.15
typing_extensions==4.10.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
